[util-iso] make intel_ucode.img optional

+++ lib/util-iso.sh
@@ -348,8 +348,10 @@ make_image_boot() {
 		copy_initcpio "${path}"
 		gen_boot_image "${path}"
 		mv ${path}/boot/${iso_name}.img ${path_iso}/${arch}/${iso_name}.img
-		cp ${path}/boot/intel-ucode.img ${path_iso}/intel_ucode.img
-		cp ${path}/usr/share/licenses/intel-ucode/LICENSE ${path_iso}/intel_ucode.LICENSE
+		if [[ -f ${path}/boot/intel-ucode.img ]]; then
+			cp ${path}/boot/intel-ucode.img ${path_iso}/intel_ucode.img
+			cp ${path}/usr/share/licenses/intel-ucode/LICENSE ${path_iso}/intel_ucode.LICENSE
+		fi
 		umount_image_handler
 		rm -R ${path}
 		: > ${work_dir}/build.${FUNCNAME}
